+++ crates/teloxide-core/src/adaptors/erased.rs
@@ -277,6 +277,8 @@ where
         get_user_chat_boosts,
         set_my_commands,
         get_business_connection,
+        get_managed_bot_token,
+        replace_managed_bot_token,
         get_my_commands,
         set_my_name,
         get_my_name,
@@ -292,6 +294,7 @@ where
         answer_inline_query,
         answer_web_app_query,
         save_prepared_inline_message,
+        save_prepared_keyboard_button,
         edit_message_text,
         edit_message_text_inline,
         edit_message_caption,
@@ -869,6 +872,16 @@ trait ErasableRequester<'a> {
         business_connection_id: BusinessConnectionId,
     ) -> ErasedRequest<'a, GetBusinessConnection, Self::Err>;
 
+    fn get_managed_bot_token(
+        &self,
+        user_id: UserId,
+    ) -> ErasedRequest<'a, GetManagedBotToken, Self::Err>;
+
+    fn replace_managed_bot_token(
+        &self,
+        user_id: UserId,
+    ) -> ErasedRequest<'a, ReplaceManagedBotToken, Self::Err>;
+
     fn get_my_commands(&self) -> ErasedRequest<'a, GetMyCommands, Self::Err>;
 
     fn set_my_name(&self) -> ErasedRequest<'a, SetMyName, Self::Err>;
@@ -914,6 +927,12 @@ trait ErasableRequester<'a> {
         user_id: UserId,
         result: InlineQueryResult,
     ) -> ErasedRequest<'a, SavePreparedInlineMessage, Self::Err>;
+
+    fn save_prepared_keyboard_button(
+        &self,
+        user_id: UserId,
+        button: KeyboardButton,
+    ) -> ErasedRequest<'a, SavePreparedKeyboardButton, Self::Err>;
 
     fn edit_message_text(
         &self,
@@ -2010,6 +2029,20 @@ where
         Requester::get_business_connection(self, business_connection_id).erase()
     }
 
+    fn get_managed_bot_token(
+        &self,
+        user_id: UserId,
+    ) -> ErasedRequest<'a, GetManagedBotToken, Self::Err> {
+        Requester::get_managed_bot_token(self, user_id).erase()
+    }
+
+    fn replace_managed_bot_token(
+        &self,
+        user_id: UserId,
+    ) -> ErasedRequest<'a, ReplaceManagedBotToken, Self::Err> {
+        Requester::replace_managed_bot_token(self, user_id).erase()
+    }
+
     fn get_my_commands(&self) -> ErasedRequest<'a, GetMyCommands, Self::Err> {
         Requester::get_my_commands(self).erase()
     }
@@ -2084,6 +2117,14 @@ where
         result: InlineQueryResult,
     ) -> ErasedRequest<'a, SavePreparedInlineMessage, Self::Err> {
         Requester::save_prepared_inline_message(self, user_id, result).erase()
+    }
+
+    fn save_prepared_keyboard_button(
+        &self,
+        user_id: UserId,
+        button: KeyboardButton,
+    ) -> ErasedRequest<'a, SavePreparedKeyboardButton, Self::Err> {
+        Requester::save_prepared_keyboard_button(self, user_id, button).erase()
     }
 
     fn edit_message_text(