+++ crates/teloxide-core/src/types/allowed_update.rs
@@ -28,4 +28,5 @@ pub enum AllowedUpdate {
     ChatJoinRequest,
     ChatBoost,
     RemovedChatBoost,
+    ManagedBot,
 }