+++ crates/teloxide-core/src/types/poll_answer.rs
@@ -21,6 +21,11 @@ pub struct PollAnswer {
     ///
     /// May be empty if the user retracted their vote.
     pub option_ids: Vec<u8>,
+
+    /// Persistent identifiers of answer options, chosen by the user.
+    ///
+    /// May be empty if the user retracted their vote.
+    pub option_persistent_ids: Vec<String>,
 }
 
 /// These fields `chat` and `user` from the original [`PollAnswer`] should be
@@ -51,7 +56,8 @@ mod tests {
         let json = r#"{
             "poll_id": "POLL_ID",
             "user": {"id": 42,"is_bot": false,"first_name": "blah"},
-            "option_ids": []
+            "option_ids": [],
+            "option_persistent_ids": []
         }"#;
 
         let poll_answer: PollAnswer = serde_json::from_str(json).unwrap();
@@ -68,7 +74,8 @@ mod tests {
                 "title": "a",
                 "type": "group"
             },
-            "option_ids": []
+            "option_ids": [],
+            "option_persistent_ids": []
         }"#;
 
         let poll_answer: PollAnswer = serde_json::from_str(json).unwrap();
@@ -85,7 +92,8 @@ mod tests {
                 "type": "group"
             },
             "user": {"id": 136817688,"is_bot": true,"first_name": "Channel_Bot"},
-            "option_ids": []
+            "option_ids": [],
+            "option_persistent_ids": []
         }"#;
 
         let poll_answer: PollAnswer = serde_json::from_str(json).unwrap();