+++ crates/teloxide-core/src/types/update.rs
@@ -4,9 +4,9 @@ use serde_json::Value;
 
 use crate::types::{
     BusinessConnection, BusinessMessagesDeleted, CallbackQuery, Chat, ChatBoostRemoved,
-    ChatBoostUpdated, ChatJoinRequest, ChatMemberUpdated, ChosenInlineResult, InlineQuery, Message,
-    MessageReactionCountUpdated, MessageReactionUpdated, PaidMediaPurchased, Poll, PollAnswer,
-    PreCheckoutQuery, ShippingQuery, User,
+    ChatBoostUpdated, ChatJoinRequest, ChatMemberUpdated, ChosenInlineResult, InlineQuery,
+    ManagedBotUpdated, Message, MessageReactionCountUpdated, MessageReactionUpdated,
+    PaidMediaPurchased, Poll, PollAnswer, PreCheckoutQuery, ShippingQuery, User,
 };
 
 /// This [object] represents an incoming update.
@@ -153,6 +153,10 @@ pub enum UpdateKind {
     /// chat to receive these updates.
     RemovedChatBoost(ChatBoostRemoved),
 
+    /// A new bot was created to be managed by the bot, or token or owner of a
+    /// managed bot was changed.
+    ManagedBot(ManagedBotUpdated),
+
     /// An error that happened during deserialization.
     ///
     /// This allows `teloxide` to continue working even if telegram adds a new
@@ -196,6 +200,7 @@ impl Update {
             ChatJoinRequest(r) => &r.from,
             ChatBoost(b) => return b.boost.source.user(),
             RemovedChatBoost(b) => return b.source.user(),
+            ManagedBot(m) => return Some(&m.user),
 
             MessageReactionCount(_) | DeletedBusinessMessages(_) | Poll(_) | Error(_) => {
                 return None
@@ -287,6 +292,7 @@ impl Update {
                 }
                 i5(empty())
             }
+            UpdateKind::ManagedBot(m) => i1(once(&m.user)),
 
             UpdateKind::ChatJoinRequest(_)
             | UpdateKind::MessageReactionCount(_)
@@ -316,6 +322,7 @@ impl Update {
             MessageReactionCount(r) => &r.chat,
             ChatBoost(b) => &b.chat,
             RemovedChatBoost(b) => &b.chat,
+            ManagedBot(_) => return None,
             DeletedBusinessMessages(m) => &m.chat,
 
             InlineQuery(_)
@@ -452,6 +459,9 @@ impl<'de> Deserialize<'de> for UpdateKind {
                             .next_value::<ChatBoostRemoved>()
                             .ok()
                             .map(UpdateKind::RemovedChatBoost),
+                        "managed_bot" => {
+                            map.next_value::<ManagedBotUpdated>().ok().map(UpdateKind::ManagedBot)
+                        }
                         _ => Some(empty_error()),
                     })
                     .unwrap_or_else(empty_error);
@@ -526,6 +536,7 @@ impl Serialize for UpdateKind {
             UpdateKind::RemovedChatBoost(v) => {
                 s.serialize_newtype_variant(name, 22, "removed_chat_boost", v)
             }
+            UpdateKind::ManagedBot(v) => s.serialize_newtype_variant(name, 23, "managed_bot", v),
             UpdateKind::Error(v) => v.serialize(s),
         }
     }
@@ -594,6 +605,7 @@ mod test {
                     is_premium: false,
                     has_topics_enabled: false,
                     allows_users_to_create_topics: false,
+                    can_manage_bots: false,
                     added_to_attachment_menu: false,
                 }),
                 sender_chat: None,
@@ -618,6 +630,7 @@ mod test {
                     external_reply: None,
                     quote: None,
                     reply_to_checklist_task_id: None,
+                    reply_to_poll_option_id: None,
                     reply_to_story: None,
                     sender_boost_count: None,
                     edit_date: None,
@@ -956,6 +969,7 @@ mod test {
                     is_premium: true,
                     has_topics_enabled: false,
                     allows_users_to_create_topics: false,
+                    can_manage_bots: false,
                     added_to_attachment_menu: false,
                 }),
                 date: DateTime::from_timestamp(1721306082, 0).unwrap(),
@@ -1138,6 +1152,7 @@ mod test {
                             is_premium: true,
                             has_topics_enabled: false,
                             allows_users_to_create_topics: false,
+                            can_manage_bots: false,
                             added_to_attachment_menu: false,
                         },
                     }),
@@ -1200,6 +1215,7 @@ mod test {
                         is_premium: true,
                         has_topics_enabled: false,
                         allows_users_to_create_topics: false,
+                        can_manage_bots: false,
                         added_to_attachment_menu: false,
                     },
                 }),