+++ crates/teloxide/src/utils/html.rs
@@ -242,6 +242,7 @@ mod tests {
             is_premium: false,
             has_topics_enabled: false,
             allows_users_to_create_topics: false,
+            can_manage_bots: false,
             added_to_attachment_menu: false,
         };
         assert_eq!(user_mention_or_link(&user_with_username), "@abcd");
@@ -255,6 +256,7 @@ mod tests {
             is_premium: false,
             has_topics_enabled: false,
             allows_users_to_create_topics: false,
+            can_manage_bots: false,
             added_to_attachment_menu: false,
         };
         assert_eq!(